requirements.txt
opencv-python>=4.9
scikit-image>=0.23
pillow
pytesseract
python-telegram-bot>=21
langchain
python-dotenv